Refactor: Split Descriptions into MD Files
Further file splits to make editing documentation and editing pages easier. Trees are now defined as separate json files. The elements of the trees match the names of md  files which are used for the description of those elements allowing for user friendly editing of documentation that aligns with the current methods used in the Studio-Whip repo.