Added Graphs
We are using D3.js to draw graphs on the website. To start we have a basic interactive graph of the /gui_framework/renderer/ files that show how they relate to one another. You can click a node to highlight it and its connecting noodles.

Doing this also required making some updates to the rust-doc-tool as we need to look for connection further than one level due to mod.rs and in some cases lib.rs being intermediary's that we want to filter out of the displayed graph without loosing the connectors.