Tree Color Coding and Graph Visual Improvments
Simplified the tree view by color coding the branches to match the graph edge colors for directionality, this way we don't have to write out both the too and from nodes to express the relation.

Improved the graph visually by making unrelated edges fade when a node is selected. Also increased edge width by one pixel for edges connected to a selected node.